Improved accuracy of nearestTValue at ends
Currently, nearestTValue snaps to 0.0 or 1.0 when getting close to the ends instead of performing the approximation. This is particularly an issue for longer Bezier curves where this can create noticeable issues in real applications.

This commit fixes this and ensures that the approximation is performed even close to the ends.